Rename create.[ch] to packet-create.[ch] and packet-key.[ch] to pgp-key.[ch].
Rationale:

* create.[ch] deals mostly with constructing pgp packets.
* packet-key.[ch] deals mostly with pgp_key_t operations,
not much pgp packet-specific things.
