+++ src/lib/key_store_kbx.c
@@ -32,7 +32,7 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-#include "create.h"
+#include "packet-create.h"
 #include "key_store.h"
 #include "key_store_pgp.h"
 #include "key_store_kbx.h"
