+++ src/lib/packet-create.c
@@ -82,7 +82,7 @@ __RCSID("$NetBSD: create.c,v 1.38 2010/11/15 08:03:39 agc Exp $");
 #include "bn.h"
 #include "rsa.h"
 #include "elgamal.h"
-#include "create.h"
+#include "packet-create.h"
 #include "key_store_pgp.h"
 #include "packet.h"
 #include "signature.h"
@@ -92,7 +92,7 @@ __RCSID("$NetBSD: create.c,v 1.38 2010/11/15 08:03:39 agc Exp $");
 #include "memory.h"
 #include "rnpdefs.h"
 #include "rnpdigest.h"
-#include "packet-key.h"
+#include "pgp-key.h"
 #include "rnpsdk.h"
 #include "ecdsa.h"
 #include "constants.h"
