+++ src/lib/packet-parse.c
@@ -82,11 +82,11 @@ __RCSID("$NetBSD: packet-parse.c,v 1.51 2012/03/05 02:20:18 christos Exp $");
 #include "bn.h"
 #include "packet-parse.h"
 #include "packet-print.h"
-#include "packet-key.h"
+#include "pgp-key.h"
 #include "key_store_pgp.h"
 #include "errors.h"
 #include "packet-show.h"
-#include "create.h"
+#include "packet-create.h"
 #include "readerwriter.h"
 #include "rnpdefs.h"
 #include "crypto.h"
