+++ src/lib/signature.c
@@ -81,7 +81,7 @@ __RCSID("$NetBSD: signature.c,v 1.34 2012/03/05 02:20:18 christos Exp $");
 #include "dsa.h"
 #include "eddsa.h"
 #include "hash.h"
-#include "create.h"
+#include "packet-create.h"
 #include "rnpsdk.h"
 #include "rsa.h"
 #include "readerwriter.h"
