+++ src/lib/writer.c
@@ -73,7 +73,7 @@ __RCSID("$NetBSD: writer.c,v 1.33 2012/03/05 02:20:18 christos Exp $");
 #endif
 
 #include "bn.h"
-#include "create.h"
+#include "packet-create.h"
 #include "writer.h"
 #include "key_store_pgp.h"
 #include "signature.h"
