+++ src/tests/cipher.c
@@ -31,7 +31,7 @@
 #include <crypto.h>
 #include <key_store_pgp.h>
 #include <packet.h>
-#include <packet-key.h>
+#include <pgp-key.h>
 #include <bn.h>
 #include <rnp.h>
 #include <ecdsa.h>
